+++ src/pages/Product/ProductDetail/ProductPurchase/PurchaseButton/PurchaseButton.js
@@ -1,13 +1,17 @@
 import React from 'react';
 import '../PurchaseButton/PutchaseButton.scss';
 
-function PurchaseButton() {
+function PurchaseButton({ postToCart, goToCart }) {
   return (
     <div className="purchaseButton">
       <div className="btnWrapper">
-        <button className="mainPurchaseBtn">구매하기</button>
+        <button className="mainPurchaseBtn" onClick={goToCart}>
+          구매하기
+        </button>
         <button className="subPurchaseBtn">찜하기</button>
-        <button className="subPurchaseBtn">장바구니</button>
+        <button className="subPurchaseBtn" onClick={postToCart}>
+          장바구니
+        </button>
       </div>
     </div>
   );
